Created CurvatureCalculator for divining curvature from point list
This functionality allows roads to be input as a series of vertices
rather than as horizontal alignments. Horizontal alignments were the
preferred input method because curvature could be read directly from the
alignment. However, even with two input methods tested, it was tedious
to create good alignments. This will allow a road to be input by points
and still have some idea of curvature. Curvature will be important
because vehicles will use curvature to influence their speed on the
road.