ColorDistribution added & tested successfully.
ColorDistributions had originally been left out of the task list for
java implementation. Upon further consideration, however, I realised
that when the simulator is running we will want to have among its output
the color that the simulated vehicle is actually painted. The viewer may
want to override that color with some evaluation-based color, but a base
color must be known in the absense of such override.