Update minimum distribution count & write tests
In the case of connector maximum positioning distances, there is no
reason to require that there be any entries in that collection in a
project. This is because vehicles in the network have infinite knowledge
of their route, and if there is no reason to override that, then there
is no reason to have anything in this collection. However, the
collection (even if empty) must appear in the parameters file.