fractional distributions added, no tests yet
Fractional distributions may be normal or empirical distributions. They
are guaranteed to return values in the range [0, 1] for any probability.
A use case would be to have a vehicle's aggression parameter determine
a fractional value, and then apply that to a vehicle's physical
characteristic like maximum acceleration. These will be used later for
desired acceleration fractions and desired deceleration fractions, which
will be a driving behavior characteristic.

Still need to write tests to ensure these are coded as intended.