XSD refined and corrected
The xs:any element apparently doesn't do what I think it does. I didn't intend for 'any' to be required, just available. Since the files would not validate with xs:any in, I have removed it. If we need to add it later we can.

All previous tests pass now. Next, need to write tests for the new additions.